+++ src/pycont/Types.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from dataclasses import dataclass, field
-from typing import List, Literal, Dict, Optional
+from typing import List, Literal, Dict, Optional, Tuple
 
 EventKind = Literal["SP", "LP", "BP", "HB", "DSFLOOR", "MAXSTEPS", "PARAM_MIN", "PARAM_MAX"]
 
@@ -24,6 +24,8 @@ class Branch:
 	stable: Optional[bool]
 	info: Optional[Dict] = field(default_factory=dict)
 
+	_tentativePoints : List[Tuple[np.ndarray, float]] = field(default_factory=list, init=False, repr=False)
+
 	def __init__(self, id : int, n_steps : int, u0 : np.ndarray, p0 : float):
 		M = len(u0)
 		self.id = id
@@ -31,12 +33,22 @@ class Branch:
 		self.p_path = np.zeros(n_steps+1); self.p_path[0] = p0
 		self.s_path = np.zeros(n_steps+1)
 		self._index = 1
+		self._tentativePoints = []
 
 	def addPoint(self, x : np.ndarray, s : float):
 		self.u_path[self._index, :] = x[0:-1]
 		self.p_path[self._index] = x[-1]
 		self.s_path[self._index] = s
 		self._index += 1
+
+	def addPointTentative(self, x : np.ndarray, s : float):
+		self._tentativePoints.append((np.copy(x), s))
+
+	def commit(self):
+		for point in self._tentativePoints:
+			self.addPoint(point[0], point[1])
+		self._tentativePoints.clear()
+		return self
 
 	def trim(self):
 		self.u_path = self.u_path[0:self._index,:]
@@ -48,10 +60,3 @@ class Branch:
 class ContinuationResult:
     branches: List[Branch] = field(default_factory=list)
     events: List[Event] = field(default_factory=list)
-
-class PyContError(Exception):
-    """Base exception for PyCont-Lite."""
-
-class InputError(PyContError, ValueError):
-    """Invalid user input or options."""
-    pass