+++ src/pycont/__init__.py
@@ -2,7 +2,7 @@ from .continuation import pseudoArclengthContinuation as arclengthContinuation
 from .plotting import plotBifurcationDiagram
 from .Logger import Verbosity
 
-__version__ = "0.3.1"
+__version__ = "0.5.0"
 
 __all__ = [
     "arclengthContinuation",